runnning but incorrect values
